Add controls to UI
Add Bills button doesn't actually perform an action. It is used to take focus away from the last input field so the 'change' event is fired in the case that the user does not do so themselves by pressing **tab**, **enter**, entering a new field, or tapping out of the fields.

Add clear button to reset all values to zero